toma

+++ BarrosoVictor_MaymoDidac_Rockets/src/com/rockets/app/application/dto/PropellantDTO.java
@@ -5,14 +5,14 @@ import com.rockets.app.utilities.InvalidParamException;
 
 
 public class PropellantDTO {
-	private Double maxAcceleration;
-	public PropellantDTO(Double maxAcceleration) throws InvalidParamException {
+	private int maxAcceleration;
+	public PropellantDTO(int maxAcceleration) throws InvalidParamException {
 		this.maxAcceleration = maxAcceleration;
 	}
 	public PropellantDTO(Propellant propellant) {
 		this.maxAcceleration = propellant.getMaxAcceleration();
 	}
-	public double getMaxAcceleration() throws InvalidParamException {
+	public int getMaxAcceleration() throws InvalidParamException {
 		if (maxAcceleration <= 0) {
 			throw new InvalidParamException();
 		}
